% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLAGS.R
\name{makeLAGS}
\alias{makeLAGS}
\title{Lag multiple variables}
\usage{
makeLAGS(data, panel, nlags, x = TRUE, y = TRUE)
}
\arguments{
\item{data}{data of class \code{data.frame}}

\item{panel}{the panel or cross-section variable name in the provided data}

\item{nlags}{the number of lags (e.g. 1, 2, ...)}

\item{x}{difference the indenpendent variables (defaults to TRUE)}

\item{y}{difference the indenpendent variables (defaults to TRUE)}
}
\value{
\code{data.table} with lagged variables
}
\description{
Create \emph{n}-lagged independent and dependent variables in a data frame. Assumes dependent variable names begin with "y_" and independent variable names begin with "x_". This makes it easier to lag many variables at once. If you variables are not so named, you can rename them with \code{panelGMM::renameVars}.
}
\note{
Requires \code{library(data.table)}. The function will automatically coerce the data.frame object to \code{data.table}. Note that \code{data.table} objects retain their \code{data.frame} class.
}
\examples{
# Using the hours and wages data provided by the panelGMM package:
data("hours_wages", package = "panelGMM")
# twice lag the independent variables:
hours_wages_gmm <- makeLAGS(hours_wages, id, 2, y = FALSE)
}
\author{
Lawrence R De Geest, \email{lrdegeest@gmail.com}
}
