% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameVars.R
\name{renameVars}
\alias{renameVars}
\title{Rename variables in a data frame}
\usage{
renameVars(data, yvars, xvars)
}
\arguments{
\item{data}{save csv files of each year's data to your directory. Defaults to FALSE.}

\item{yvars}{vector (\code{c()}) or single dependent variable name(s)}

\item{xvars}{vector (\code{c()}) or single independent variable name(s)}
}
\value{
\code{data.frame} with "x_" prepended to all independent variables names and "y_" prepended to all dependent variable names
}
\description{
Append prefixes "y_" and "x_" respectively to the names of dependent and indenpent variables. This facilitates differencing and lagging variables in batches using \code{panelGMM::makeDIFFS} and \code{panelGMM::makeLAGS}.
}
\examples{
# Using the hours and wages data provided by the panelGMM package:
data("hours_wages", package = "panelGMM")
hours_wages_gmm <- renameVars("lnhr", c("lnwg", "kids", "age", "agesq", "disab"), hours_wages)
}
\author{
Lawrence R De Geest, \email{lrdegeest@gmail.com}
}
