% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftCols.R
\name{shiftCols}
\alias{shiftCols}
\title{Shift columns to create lags, leads or differences}
\usage{
shiftCols(data, panel, type = c("lag", "lead", "difference"), shifts, columns)
}
\arguments{
\item{data}{data of class \code{data.frame}}

\item{panel}{the panel or cross-section variable name in the data}

\item{type}{shift type. either "lag", "lead" or "difference"}

\item{shifts}{the number of lags, leads or differences}

\item{columns}{the columns in `data` to be lagged, leaded or differenced}
}
\value{
\code{data.frame} / \code{data.table} with new variables
}
\description{
Create \emph{n}-lagged/leaded/differenced columns in a data frame with automatic names.
}
\note{
The function will automatically coerce the data.frame object to a \code{data.table} object.  (\code{data.table} objects retain their \code{data.frame} class.)
}
\examples{
# Using the hours and wages data provided by the panelGMM package:
data("hours_wages")
# shiftCols() is designed to work nicely with the pipe operator `\%>\%` from `magrittr`:
hours_wages_gmm = hours_wages \%>\%
   shiftCols(data = ., panel=id, type = "difference", shifts = 1, columns=c("lnhr", "lnwg", "kids", "age", "agesq", "disab")) \%>\%
   shiftCols(data = ., panel=id, type = "lag", shifts = 1:4, columns=c("lnhr", "lnwg", "kids", "age", "agesq", "disab"))
}
\author{
Lawrence R De Geest, \email{lrdegeest@gmail.com}
}
