% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelGMM.R
\name{panelGMM}
\alias{panelGMM}
\title{Generalized Method of Moments (GMM) for panel data}
\usage{
panelGMM(formula, panel, time, twostep = TRUE, intercept = FALSE, data)
}
\arguments{
\item{formula}{the model in standard R form with instruments (e.g., \code{y ~ x1 + x2 | z1 + z2 ...}).}

\item{panel}{the name of the panel variable in the data.}

\item{time}{the name of the time variable in the data.}

\item{twostep}{estimate a two-step GMM model. Defaults to TRUE.}

\item{intercept}{optional intercept. Defaults to FALSE.}

\item{data}{data of class \code{data.frame}/\code{tibble}/\code{data.table}.}
}
\value{
a \code{panelGMM} object.
}
\description{
Fit one-step and two-step GMM models for panel data with lagged instrumental variables. Calculates panel-robust standard errors allowing for heteroskedasticity and correlation over time.
Includes \code{summary.panelGMM} and \code{print.summary.panelGMM} methods.
}
\note{
Currently works only for balanced panels.
}
\examples{
# Using the hours and wages data provided by the panelGMM package:
data("hours_wages", package = "panelGMM")
# create lagged and differenced instruments
hours_wages_gmm = hours_wages \%>\%
   shiftCols(data = ., panel=id, type = "difference", shifts = 1, columns=c("lnhr", "lnwg", "kids", "age", "agesq", "disab")) \%>\%
   shiftCols(data = ., panel=id, type = "lag", shifts = 1:4, columns=c("lnhr", "lnwg", "kids", "age", "agesq", "disab"))
# define the model
model = lnhr_diff_1 ~ lnwg_diff_1 + kids_diff_1 + age_diff_1 + agesq_diff_1 + disab_diff_1 | kids_lag_1 + kids_lag_2 + age_lag_1 + age_lag_2 + agesq_lag_1 + agesq_lag_2 + disab_lag_1 + disab_lag_2 + lnwg_lag_2
# estimate the parameters with two-step GMM
model_results = panelGMM(model, panel = id, time = year, twostep = TRUE, data = hours_wages_gmm)
summary(model_results)
}
\author{
Lawrence R. De Geest, \email{lrdegeest@gmail.com}
}
